corregiendo problemas coordenadas

+++ model/init_db.sql
@@ -14,8 +14,8 @@ CREATE TABLE pizzerias (
     id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     address VARCHAR(255) NOT NULL,
-    latitude FLOAT NOT NULL,
-    longitude FLOAT NOT NULL
+    latitude DECIMAL(17,15) NOT NULL,
+    longitude DECIMAL(17,15) NOT NULL
 );
 
 CREATE TABLE favourites (
